reorganized js files into modules
